Split track_reads task into several steps
* We already reduced the memory required for track-reads.
* This turns select-reads into 2-level partitioning:
  * Partition from BAM inputs to merged BAM files,
    each with only the reads for a subset of ctgs.
  * Segregate the reads from each merged BAM, so
    Quiver will see only the reads for a single ctg.

The 2nd step can run in parallel. The first step is serial,
walking thru all BAM files, writing to a limited number
of merged.bam partitions.

So far, this seems to work. But the serial step needs to be
in C because pysam is too slow. (Walking through all BAM records
takes 7 hours in one example.)

https://jira.pacificbiosciences.com/browse/TAG-1502