prototype for safe haplotig extraction
Initial commit. Updated the unzip.py to use the modified scripts, and added an initial version of the scripts to chop up the primary contigs into linear regions, augment the phasing blocks and extract the haplotigs. The overlap filter is now modified to be stricter, in the sense that only overlaps which are fully phased are let through. This allows assemblies of each individual phased component fully, without risking haplotype switching. Placement will be resolved after that.

Resolved placement and clipping of the haplotigs corresponding to the linear regions. Haplotigs are currently aligned to all the linear regions using BLASR, grouped in diploid pairs, and clipped to start and end at the same position of the same linear region. Bubbles are not yet added to the mix, that is the next step.

Added a state machine to main_augment_pb.py to detect simple and complex bubbles.

Updated the state machine to separate contig regions.

Replaced the extraction of linear region FASTA sequences with the new state machine code.

Cleaned up the main_augment_pb.py a bit.

Added loading of the json files from the proto step.

Created a phase relation graph and resolving of phase aliases, and wrote all regions to a json file.

Updated the main_augment_pb so that the paths are replaced by 'haplotig' dicts of dicts, where the first key is the haplotig name, and the second level key is one of 'name', 'seq', 'path', 'edges', 'cstart', 'cend' and 'phase'.

Changed the phase annotation from tuples to a string joined with '_' to avoid conversion issues when writing/loading from a .json file. Also, fixed the coordinate issue with SV bubbles.

The graphs_to_h_tigs_2 now can integrate the diploid regions together with all the other regions from the initial assembly by breaking the linear regions in diploid blocks. A NetworkX graph is now build from the regions. The primary contig is now resolved as the shortest path in the graph with weights assigned similar to before. This should be done better in the future to allow maximization of the number of same aliases in the graph (via dynamic programming) and long range haplotype scaffolding.

Added the extraction of alternative phased haplotigs, now written to h_tigs_all.ctg_id.fa.

Added GFA output of haplotigs.

Prevented the crash due to dedup.

Added verbose logging, string graph to GFA output, changed the weights for extracting haplotigs to be non-negative (due to issues with NetworkX). This version has an issue when multiple smaller components of the same haplotig are present, then the shortest one will be preferred. This will be resolved in the following commits.

Fixed a bug in delineate_regions where a complex bubble would falsely be called if a new bubble was opened just where the last ended. Added a new method to map preads using Minimap2 instead of BLASR.

Minor modification to sam2m4.

Minor change to tiling_path.

Modified the haplotig extraction to be much more safe - now instead of using the nx.shortest_path between sources and sinks, we actually emply the original FALCON to do the job. It generates haplotigs as contigs. The contigs are then loaded, their phases matched and haplotigs formed. Haplotigs are then aligned just as before, but this should be much safer with respect to repeats.

There was an issue I discovered with phase propagation - it turned out that some SV bubbles were phased strangely. One branch had mixed phases of the same phasing blocks, which, when directly added to the phase relation graph, made a mess and turned everything into one giant connected components. Such strange bubbles are now detected, and not added to the phase relation graph. Additionally, for Minimap2 alignments, only one mapping per qname is now taken.

Added a few more strict filters in ovlp_filter_with_phase_strict.py.

Minor change.

Moved the parts of the generate_haplotigs_for_ctg() function which generate the phaplotig_graph, and _ctg and h_ctg, into separate functions.

Removed a stderr write.

Refactored creation of the haplotig graph (nx) and updating of the aliases and edges, so that ultimately this can be run iteratively.

Moved the utility functions to a separate script, and cleaned up the unused code, so that it's more readablae.

Code cleanup and reorganization.

Minor cleanup. This commit should be functional, before a new refactor and redesigned features.

Moved the v1 version of the prototype to mains/graphs_to_h_tigs_2_v1.py, and created a copy in mains/graphs_to_h_tigs_2.py. I modified the algorithm to actually handle potential multiple haplotigs for a phasing block. This is done by first aligning all asm-falcon/p_ctg.fa, then collecting all possible break points and breaking each haplotig. After all haplotigs are broken, all pieces for a coordinate strech are pooled. If there are != 2 seqs for a coordinate span, or if there are 2 seqs for a coordinate span but not of opposite phase, this region is ignored as a potential repeat, overhang or an unresolved collapsed region. Remaining regions are grouped as diploid groups, and added as bubbles in the list of regions. The graph is built and output.

BLASR crashes on empty input, so I added a check for that.

Mostly changes in debug output. Fixed a bug where I changed the value of ctg_id in a for loop whereas that was actually a used variable.

Fixed a crash when there were 0 edges extracted from the path.

Removed commented out code.

Slight cleanup/reordering of the code.

Code cleanup and documentation. Extracted the portion of code which extracts preads for a contig ID as a separate point script, and refactored the call to this script and Minimap2 as Bash calls in the TASK script in unzip.py.

Removed unused code.

Cleaned up main_augment_pb.py and parametrized the preads SAM file.

Minor cleanup.

Handled the complex bubbles by breaking the haplotig graph into multiple components. Handled the case when there were multiple components in the final haplotig graph to allow extraction of more than one primary contig. This is arbitrary though, because if an assembly is fully phased, then there would be two primary contigs in the output instead of one in the primary and one in the haplotigs.

Added path and edges output to the prototype. This needs to be revised, because there are edges missing between haplotigs.

Fixed the conversion from TilingPathEdge to split lines form to be correct in all cases. Unzip task script now generates the all_h_ctg_ids needed for fc_quiver.

Removed an unused script.

Removed some unused code.

Minor code cleanup.