For polishing, keep all subreads for any mapped zmw
Warn if 2 subreads from the same zmw map to different ctgs.
We used to support that, but now we pick one arbitrarily.
That is a rare case. Mainly it happens when we used
all subreads for assembly AND we split a subread into 2 contigs.

Note: If we have only 1 subread for a zmw and we happen to split it into
2 contigs, we must have mapped it to only 1 ctg. So I am not sure
this problem ever actually occurs. So maybe we should consider this
an error and crash, rather than logging a warning deep in the task
stderr.

TAG-3437